fix: resolve SwiftLint violations in test files
- Fix force unwrapping by using guard let and Data(string.utf8) pattern
- Fix force try by converting helper methods to throw
- Fix force cast by using guard let with as? conditional cast
- Fix identifier names (h1/h2 -> header1/header2, i -> index, s1/s2 -> section1/section2)
- Fix function parameter count by using BitmapHeaderParams struct
- Split large files to stay under 400 lines:
  - PETests.swift -> PETests.swift + PEResourceTests.swift + PETestHelpers.swift
  - ExtensionTests.swift -> ExtensionTests.swift + FileExtensionTests.swift
  - SettingsTests.swift (trimmed duplicate tests)
- Fix trailing newlines via SwiftFormat
